Change cache.extra to be a per chunk option
Changes cache.extra to the be a per chunk option rather than a package option.
This allows you to set a particular chunk's cache to be invalidated if the
modification time of the file is updated for instance, among numerous other
uses.  The previous behavior can still be used by specifying the cache.extra
chunk option in the first chunk to make it the default, as with the regular
cache option. Fixes #404