Bears!!
Greg @gfinak said this would be my last GIT commit message if I push to Github in Alaska, since I'll be there for vacation for the next week. Let's see if he is right. I hope not, of course.

Anyway, this is knitr v1.4, and I start to pray CRAN masters won't bite me. Well, it is way better for them to bite me rather than bears biting me in Alaska.