I cannot just use rmarkdown::pandoc_available() when registering vignette engines, because R CMD check will give the following warning:
* checking S3 generic/method consistency ... WARNING
Error in match(x, table, nomatch = 0L) :
  'match' requires vector arguments
Calls: <Anonymous> -> checkArgs -> %in% -> match
Execution halted
See section ‘Generic functions and methods’ of the ‘Writing R
Extensions’ manual.

I do not really understand what it means. Perhaps it is a bug in R. It is too deep for me to investigate, but I think I know the problem is due to the fact that I loaded the namespace of rmarkdown, which will also load digest and htmltools, then there seems to be some S3 mess that R does not like. I wrote this short version of pandoc_available() by myself, which fixed the problem.