No warning if chunk with child has only whitespace
This warning is useful to make sure authors are aware that code in
the chunk will not be run. But because pure whitespace is not code,
there is no concern.

This commit has the extra benefit of increasing compatibility with
LyX's chunk inset. An empty chunk inset in LyX consists of a blank
line, and would thus trigger the warning without this commit.