Added clarifying note about fig.path behavior (#1338)
Added a note which should help clarify the behavior of the `fig.path` chunk option and should help make it easier to see how to change the figure output directory to a relative location. In particular, it is useful to indicate that 'output/dir' and 'output/dir/' will behave very differently when passed to `fig.path`. (e.g. see http://stackoverflow.com/questions/21582402/knitr-how-to-set-a-figure-path-in-knit2html-without-using-setwd)

ported from 7529d8ae691335764cd1558bbab904250ba01ca2