I added the copyright notice recommended by GNU.
To the README.md, I added the copyright notice recommended at
this webpage.
https://www.gnu.org/licenses/gpl-howto.html

The recommendation in fact is to add the copyright notice to all
applicable files, but I decided to start with something less involved.

As noted in a prior commit, you need to confirm that GPLv3+ is the
correct version or to change the version accordingly.