Allow setting only convert.opts for tikz engine
The *convert.opts* field could not be given as the only field
to *engine.opts* when using the tikz engine. The default value
for *convert* was not chosen if *convert.opts* was set, as R used
partial matching for `engine.opts$convert`. Using the `[[` extractor to
select the list element fixes this.