utils-vignettes: Don't disable vtangle if .Rout.save exists
When inside CMD check, vtangle() returns an empty file to avoid
unnecessary code execution (that already ran as part of the weave), and
prevent vignettes that don't tangle correctly breaking check output.
However, this output is needed if you're trying to compare tangled
output to an ".Rout.save".

This whole code block is nearly redunant, CMD check will perform
similar checks and not tangle, but only if
_R_CHECK_VIGNETTES_SKIP_RUN_MAYBE_ is enabled, which it currently isn't
by default.