Add quarto support for #| (#2134)
The approach taken here had the goal of not touching the original spin code, so that this only impa
cts `format = 'qmd'` and did so as simply as possible.

To do this, I find the any code chunk that *starts* with `#|` and prepend a `# %%` for the previous
 line. This prevents converting each option passed with `#|` from creating a new chunk.

Perhaps this could be integrated more deeply with how knitr::spin works, but this is the simplest s
olution.